Application/requirements.txt
uvicorn == 0.19.0
fastapi == 0.85.1
python-dotenv == 0.21.0
requests == 2.28.1
pytest == 7.1.3
aiohttp == 3.8.3
azure-core == 1.26.1 
azure.cosmos == 4.3.0